+++ src/dreamstate/test-utils.ts
@@ -8,8 +8,10 @@ export * from "@/dreamstate/test-utils/provision/mockScopeProvider";
 export * from "@/dreamstate/test-utils/services/mockManagers";
 export * from "@/dreamstate/test-utils/services/mockManager";
 
+export * from "@/dreamstate/test-utils/registry/mockRegistry";
 export * from "@/dreamstate/test-utils/registry/getCurrent";
 export * from "@/dreamstate/test-utils/registry/mockScope";
+export * from "@/dreamstate/test-utils/registry/mockManagerInitialContext";
 
 export * from "@/dreamstate/test-utils/utils/getReactConsumer";
 export * from "@/dreamstate/test-utils/utils/getReactProvider";