+++ src/dreamstate/test-utils/registry/mockRegistry.test.ts
@@ -1,0 +1,8 @@
+import { createRegistry } from "@/dreamstate/core/scoping/registry/createRegistry";
+import { mockRegistry } from "@/dreamstate/test-utils/registry/mockRegistry";
+
+describe("mockRegistry method", () => {
+  it("Should be same as registry creation method", () => {
+    expect(mockRegistry()).toStrictEqual(createRegistry());
+  });
+});