+++ src/dreamstate/test-utils/registry/mockScope.test.ts
@@ -1,6 +1,8 @@
 import { createRegistry, IRegistry } from "@/dreamstate/core/scoping/registry/createRegistry";
 import { IScopeContext } from "@/dreamstate/core/scoping/ScopeContext";
+import { mockManagerInitialContext } from "@/dreamstate/test-utils/registry/mockManagerInitialContext";
 import { mockScope } from "@/dreamstate/test-utils/registry/mockScope";
+import { TestManager } from "@/fixtures";
 
 /**
  * todo: Add lifecycle toggling tests.
@@ -23,8 +25,25 @@ describe("mockScope test util", () => {
 
   it("Should properly inject provided registry", () => {
     const registry: IRegistry = createRegistry();
-    const scope: IScopeContext = mockScope(true, registry);
+    const scope: IScopeContext = mockScope({ isLifecycleDisabled: true }, registry);
 
     expect(scope.INTERNAL.REGISTRY).toBe(registry);
+  });
+
+  it("Should use registering with initial context", () => {
+    const scope: IScopeContext = mockScope(
+      {
+        isLifecycleDisabled: true,
+        applyInitialContexts: [ mockManagerInitialContext(TestManager, { first: "some-secret-param" }) ]
+      }
+    );
+
+    scope.INTERNAL.registerService(TestManager, {});
+
+    expect(scope.INTERNAL.REGISTRY.CONTEXT_STATES_REGISTRY.get(TestManager)).toStrictEqual({
+      first: "some-secret-param",
+      second: 2,
+      third: false
+    });
   });
 });